virtualized table, column path, checkbox, additional info
