+++ app/package.json
@@ -27,6 +27,7 @@
     "react-number-format": "^5.3.1",
     "react-router": "^6.8.1",
     "react-router-dom": "^6.8.1",
+    "react-virtuoso": "^4.7.11",
     "react-zoom-pan-pinch": "^3.3.0",
     "simplebar-core": "^1.2.4",
     "simplebar-react": "^3.2.4",
