+++ app/src/settings_tabs/Search.css
@@ -2,4 +2,5 @@
   display: flex;
   align-items: center;
   justify-content: center;
+  margin-top: 2vh;
 }
